+++ src/v1/database/factor.js
@@ -4,6 +4,7 @@ const { addRoznamcha } = require("./roznamcha");
 const { changeExistance } = require("./drug");
 const Sentry = require("../../log");
 const { ObjectId } = require("mongoose").Types;
+const { FactorTypeEnum, PaymentTypeEnum } = require("../utils/enum");
 const getFactors = async () => {
   const pipline = [
     {
@@ -82,12 +83,37 @@ const addFactor = async (
 };
 const deleteFactor = async (i18n, id) => {
   try {
-    const isDeletedFactor = await Factor.findByIdAndRemove(id);
-    if (isDeletedFactor) {
-      return { message: i18n.__("factor_deleted_successfully") };
-    } else {
-      return { message: i18n.__("failed_to_delete_factor") };
+    const { amount, customer, factorType, paymentType } = await Factor.findById(
+      id
+    );
+    const { balance } = await Customer.findById({ _id: customer });
+    if (
+      factorType == FactorTypeEnum.BUY &&
+      paymentType == PaymentTypeEnum.NO_CASH
+    ) {
+      const updatedCustomer = await Customer.findOneAndUpdate(
+        { _id: customer },
+        { balance: balance - amount }
+      );
+      if (updatedCustomer) {
+        const isDeletedFactor = await Factor.findByIdAndRemove(id);
+        return { message: i18n.__("factor_deleted_successfully") };
+      }
     }
+    if (
+      factorType == FactorTypeEnum.SELL &&
+      paymentType == PaymentTypeEnum.NO_CASH
+    ) {
+      const updatedCustomer = await Customer.findOneAndUpdate(
+        { _id: customer },
+        { balance: balance + amount }
+      );
+      if (updatedCustomer) {
+        const isDeletedFactor = await Factor.findByIdAndRemove(id);
+        return { message: i18n.__("factor_deleted_successfully") };
+      }
+    }
+    return { message: i18n.__("failed_to_delete_factor") };
   } catch (error) {
     Sentry.captureException(error);
     throw error;
